refactor: move AppDatastoreConfig model to data layer
